Refactor Travis-specific code into a pluggable class
